refactor: consolidate economy test Game builders into fixture_builders (Refs #4108)
Introduce lib/src/fixture_builders/game_builders.dart as the canonical
owner of inline Game(...) construction and extend repo.economy_test_core_fixtures_shared
to enforce the same rule on colonizethis_economy_test_support/lib.
