+++ SPEC/program/repo-lint.md
@@ -230,7 +230,7 @@
 | `tool/check_app_test_no_partn_double_suffix.dart` | Full-tree gate: forbid `*_test_partN_test.dart` double-suffix split fragments; require `<family>_partN_test.dart` (Refs #3730, #4035); rule `repo.app_test_no_partn_double_suffix` |
 | `tool/check_economy_bid_quantity_cap_shared.dart` | Require `trade_order_suggester.dart` to delegate bid quantity capping to `capBidQuantityForBudgets` (phase-7 home `treasury_bid_caps.dart`, re-exported via the `treasury_bid_budget.dart` barrel); forbid inline `remainingTreasuryBudget ~/ unitPrice` elsewhere under economy lib (Refs #3836, #4049); rule `repo.economy_bid_quantity_cap_shared` |
 | `tool/check_economy_test_no_src_imports.dart` | Forbid `package:colonizethis_economy/src/` imports under `packages/colonizethis_economy/test/**`; tests must use the public barrel (Refs #3836); rule `repo.economy_test_no_src_imports` |
-| `tool/check_economy_test_core_fixtures_shared.dart` | Forbid inline `Game(` construction under `packages/colonizethis_economy/test/**`; use shared builders from `colonizethis_economy_test_support` (Refs #3836, #3856); rule `repo.economy_test_core_fixtures_shared` |
+| `tool/check_economy_test_core_fixtures_shared.dart` | Forbid inline `Game(` construction under `packages/colonizethis_economy/test/**` and `packages/colonizethis_economy_test_support/lib/**` except `lib/src/fixture_builders/**`; use shared builders from `colonizethis_economy_test_support` (Refs #3836, #3856, #4108); rule `repo.economy_test_core_fixtures_shared` |
 | `tool/check_economy_deal_matcher_treasury_shared.dart` | Require `deal_matcher_session.dart` to call `maxAffordableBidQuantity` / `decrementTreasuryForFill` from the `treasury_bid_budget.dart` barrel (phase-7 home `treasury_bid_caps.dart`); forbid `_maxAffordableQuantity` / `_decrementTreasury` outside the shared module (Refs #3856 / #3979, #4049); rule `repo.economy_deal_matcher_treasury_budget_shared` |
 | `tool/check_app_no_duplicate_helpers.dart` | AST scan that pins **canonical helper symbols** extracted for #2180 (e.g. `eraRoman`, `techCategoryLabelL10n`, `commodityDisplayName`, the `trainDialog*` set) to one canonical file under `app/lib/**` and forbids reappearance of removed helpers — the previous private duplicates (`_eraRoman`, `_categoryLabel`, `_categoryLabelL10n`, `_commodityDisplayName`) and the deleted public wrapper `unitsPanelRegionLabel` (replaced by `regionDisplayLabel`, #3279) — anywhere as top-level functions or class methods (GitHub #2180, #3279); rule `repo.app_no_duplicate_helpers` |
 | `tool/check_app_test_no_debug_init.dart` | AST scan that fails when any `app/test/**/*.dart` file invokes the expensive debug map generator `getDebugInitGameResult()` (runs `runInitGame(GameSetupConfig.defaultConfig)`, ~7-11s per call paid once per `flutter test` isolate) outside an in-script allowlist. The allowlist is the migration backlog: every `app/test/**` file that still calls the helper; it must only shrink as panel suites move to the shared lightweight fixtures in `app/test/support/panel_test_fixtures.dart` or a committed serialized fixture, and new entries require justification for a genuine generated-map/topology dependency. Comment-only mentions of the symbol are ignored (AST method-invocation match). The shrink-only invariant is enforced: a **stale allowlist entry** (file missing, or migrated so it no longer invokes the helper) also fails the check and must be removed (Refs #3656); rule `repo.app_test_no_debug_init` |
