+++ packages/colonizethis_economy_test_support/lib/src/boycott_blocked_commodities_test_support.dart
@@ -1,43 +1,15 @@
 // dart format off
-// Shared fixtures for `boycottedColonySellableCommodityIds` tests (Refs #3758
-// S7/R12; #3831 Phase 4 hoisting from near-cap economy test files).
+// Shared fixtures for `boycottedColonySellableCommodityIds` tests (Refs #3758 S7/R12; #3831; #4108).
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
 import 'extraction_fixture_support.dart';
-/// New-World region id used by boycott colony-tribe fixtures.
+export 'fixture_builders/game_builders.dart' show gameWithColonyTribeBoycottTest;
 const boycottColonyTribeNwRegionId = 'newWorld';
-/// Prefixed province id for the single-tile colony Tribe in boycott fixtures.
 const boycottColonyTribeProvinceId = 'newWorld|t1';
-/// Builds a minimal game with a one-tile New-World tribe `t1` whose developed
-/// tile produces furs so `computeNonGreatPowerAutoOffers` emits a furs offer.
-Game gameWithColonyTribeBoycottTest({List<ColonyState> colonyStates = const [], List<BoycottState> boycottStates = const []}) {
-  final tileState = TileMapState().setImprovement('newWorld|t1|0|0', 1).setRoadLevel('newWorld|t1|0|0', 1);
-  return Game(
-    id: 'g_boycott_test',
-    worldState: WorldState(
-      turnState: TurnState(turnNumber: 1, phase: TurnPhase.orders),
-      oldWorld: const RegionData(),
-      newWorld: RegionData(
-        provinces: [Province(id: boycottColonyTribeProvinceId, regionId: boycottColonyTribeNwRegionId, ownerId: 't1', townDevelopmentLevel: 1)],
-      ),
-      tileState: tileState,
-    ),
-    players: [
-      Player(id: 'gpA', displayName: 'Aragon', isHuman: false),
-      Player(id: 'gpC', displayName: 'Castile', isHuman: false),
-    ],
-    tribes: [testTribe()],
-    colonyStates: colonyStates,
-    boycottStates: boycottStates,
-  );
-}
-/// Tile maps for [gameWithColonyTribeBoycottTest] (single furs tile in `t1`).
 Map<String, TileMapResult> tileMapsForBoycottColonyTribeTest() => {
   boycottColonyTribeNwRegionId: nonGpProvMap(boycottColonyTribeProvinceId, const [
     [Resource.furs],
   ]),
 };
-/// Minimal topology for boycott colony-tribe fixtures.
 MapTopology topologyForBoycottColonyTribeTest() => const MapTopology(
   nodes: [TopologyNode(id: 't1', regionId: boycottColonyTribeNwRegionId, type: TopologyNodeType.province)],
   edges: [],
