+++ packages/colonizethis_economy_test_support/lib/src/core_economy_test_support.dart
@@ -1,8 +1,10 @@
 // dart format off
-// Shared builders for core economy unit tests (Refs #3836).
-import 'package:colonizethis_data/colonizethis_data.dart';
+// Shared builders for core economy unit tests (Refs #3836, #4108).
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/game_test_fixtures.dart';
+
+export 'fixture_builders/game_builders.dart';
+
 /// Minimal player fixture for build-cost and worker-economy suites.
 Player corePlayer({String id = 'p1', String displayName = 'P', bool isHuman = true, int treasury = 0, Stockpile stockpile = const Stockpile(), Map<String, bool> techUnlocked = const {}}) {
   return Player(id: id, displayName: displayName, isHuman: isHuman, treasury: treasury, stockpile: stockpile, techUnlocked: techUnlocked);
@@ -18,42 +20,6 @@ Stockpile stockpileWithDeltas(Map<CommodityId, int> deltas) {
     stockpile = stockpile.applyDelta(key, value);
   }
   return stockpile;
-}
-/// Two-player game skeleton for trade-interception tests.
-Game tradeInterceptionGame({String id = 'g1', List<Fleet> fleets = const [], List<DiplomacyRelation> diplomacyRelations = const [], List<Player>? players, RelationState defaultRelation = RelationState.atPeace}) {
-  final resolvedPlayers = players ?? const [Player(id: 'p1', displayName: 'A', isHuman: true), Player(id: 'p2', displayName: 'B', isHuman: true)];
-  final resolvedRelations = diplomacyRelations.isNotEmpty ? diplomacyRelations : [DiplomacyRelation(factionId1: 'p1', factionId2: 'p2', state: defaultRelation)];
-  return Game(
-    id: id,
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: const RegionData(),
-      newWorld: const RegionData(),
-      fleets: fleets,
-    ),
-    players: resolvedPlayers,
-    diplomacyRelations: resolvedRelations,
-  );
-}
-/// Enemy patrol fleet with optional privateering tech on player p2.
-Game tradeInterceptionPrivateeringGame({required bool enemyHasPrivateering, String gameId = 'g1'}) {
-  return Game(
-    id: gameId,
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: const RegionData(),
-      newWorld: const RegionData(),
-      fleets: [
-        Fleet(id: 'enemy', ownerId: 'p2', seaZoneId: 'sea1', regionId: 'oldWorld', shipTypeIds: const ['sloop'], mission: FleetMission.patrol),
-        Fleet(id: 'mine', ownerId: 'p1', seaZoneId: 'sea1', regionId: 'oldWorld', shipTypeIds: const ['fluyte']),
-      ],
-    ),
-    players: [
-      const Player(id: 'p1', displayName: 'A', isHuman: true),
-      Player(id: 'p2', displayName: 'B', isHuman: false, techUnlocked: enemyHasPrivateering ? const {kTechIdPrivateeringCompanies: true} : const {}),
-    ],
-    diplomacyRelations: const [DiplomacyRelation(factionId1: 'p1', factionId2: 'p2', state: RelationState.atWar)],
-  );
 }
 var _tradeInterceptionFleetSeq = 0;
 /// Deterministic fleet factory for trade-interception scan tests.
@@ -72,28 +38,4 @@ Game minimalEconomyGame({String id = 'g1', int turnNumber = 0, List<Player>? pla
   players: players ?? const [Player(id: 'h1', displayName: 'Human', isHuman: true)],
   fleets: fleets ?? const [],
 );
-/// Two-player game skeleton for extraction application tests.
-Game minimalTwoPlayerGame({String id = 'g1', int turnNumber = 0, List<Player>? players}) {
-  return Game(
-    id: id,
-    worldState: WorldState(
-      turnState: TurnState(phase: TurnPhase.orders, turnNumber: turnNumber),
-      oldWorld: const RegionData(),
-      newWorld: const RegionData(),
-    ),
-    players: players ?? const [Player(id: 'p1', displayName: 'A', isHuman: true), Player(id: 'p2', displayName: 'B', isHuman: false)],
-  );
-}
-/// Single-GP game for trade-cargo-capacity suites.
-Game minimalGpGame({String id = 'g1', String playerId = 'gp1', int turnNumber = 1}) {
-  return Game(
-    id: id,
-    players: [Player(id: playerId, displayName: playerId, isHuman: false, stockpile: Stockpile.empty, treasury: 0)],
-    worldState: WorldState(
-      turnState: TurnState(phase: TurnPhase.orders, turnNumber: turnNumber),
-      oldWorld: const RegionData(),
-      newWorld: const RegionData(),
-    ),
-  );
-}
 // dart format on
